Create aws-ci-cd-pipeline.py
n this example, the codecommit, codebuild, and s3 clients are created for the respective AWS services. The source_stage, build_stage, and deploy_stage variables define the pipeline stages, which are then passed to the create_pipeline method to create the pipeline.

The canary variable defines the canary deployment strategy by putting a new revision of the deployment.
However, this is a blueprint designed for those of you who need this code in order to deploy you application.